+++ app/src/cc/arduino/contributions/libraries/ui/LibrariesIndexTableModel.java
@@ -29,8 +29,6 @@
 
 package cc.arduino.contributions.libraries.ui;
 
-import cc.arduino.contributions.DownloadableContributionBuiltInAtTheBottomComparator;
-import cc.arduino.contributions.filters.InstalledPredicate;
 import cc.arduino.contributions.libraries.ContributedLibrary;
 import cc.arduino.contributions.libraries.LibrariesIndexer;
 import cc.arduino.contributions.packages.ContributedPlatform;
@@ -38,79 +36,14 @@ import cc.arduino.contributions.ui.FilteredAbstractTableModel;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @SuppressWarnings("serial")
 public class LibrariesIndexTableModel extends FilteredAbstractTableModel<ContributedLibrary> {
 
   public final static int DESCRIPTION_COL = 0;
-
-  public static class ContributedLibraryReleases implements Comparable<ContributedLibraryReleases> {
-
-    public final String name;
-    public final List<ContributedLibrary> releases;
-    public final List<String> versions;
-
-    public ContributedLibrary selected;
-
-    public ContributedLibraryReleases(ContributedLibrary library) {
-      this.name = library.getName();
-      this.versions = new LinkedList<>();
-      this.releases = new LinkedList<>();
-      this.selected = null;
-      add(library);
-    }
-
-    public boolean shouldContain(ContributedLibrary lib) {
-      return lib.getName().equals(name);
-    }
-
-    public void add(ContributedLibrary library) {
-      releases.add(library);
-      String version = library.getParsedVersion();
-      if (version != null) {
-        versions.add(version);
-      }
-      selected = getLatest();
-    }
-
-    public ContributedLibrary getInstalled() {
-      List<ContributedLibrary> installedReleases = releases.stream().filter(new InstalledPredicate()).collect(Collectors.toList());
-      Collections.sort(installedReleases, new DownloadableContributionBuiltInAtTheBottomComparator());
-
-      if (installedReleases.isEmpty()) {
-        return null;
-      }
-
-      return installedReleases.get(0);
-    }
-
-    public ContributedLibrary getLatest() {
-      return getLatestOf(releases);
-    }
-
-    public ContributedLibrary getSelected() {
-      return selected;
-    }
-
-    public void select(ContributedLibrary value) {
-      for (ContributedLibrary plat : releases) {
-        if (plat == value) {
-          selected = plat;
-          return;
-        }
-      }
-    }
-
-    @Override
-    public int compareTo(ContributedLibraryReleases o) {
-      return name.compareToIgnoreCase(o.name);
-    }
-  }
 
   private final List<ContributedLibraryReleases> contributions = new ArrayList<>();
 
@@ -271,7 +204,7 @@ public class LibrariesIndexTableModel extends FilteredAbstractTableModel<Contrib
     contributions.clear();
     indexer.getIndex().getLibraries().forEach(this::applyFilterToLibrary);
     indexer.getInstalledLibraries().forEach(this::applyFilterToLibrary);
-    Collections.sort(contributions);
+    Collections.sort(contributions, new ContributedLibraryReleasesComparator());
   }
 
 }
